feat: switch to explicit invocation
This commit switches the plugin operation from an implicit trigger to an
explicit trigger. In other words, instead of the plugin converting links
on its own, it requires the user to explicitly invoke the operation.

The initial implementation listens on the `modify` event and the
intention was to only convert links when the document is saved. However,
it was discovered that the `modify` event can be emitted even while the
user is still editing (likely due to auto-save), which makes the
experience very jarring for the user.

As such, this commit pivots the plugin to only operate on explicit
invocation. At present, it can be invoked in the following ways:

- Command + Current Selection
- Command + Current Document
- Editor Menu + Current Selection
- Editor Menu + Current Document

To facilitate this change, the project is also updated to drop the
'Automatically' description.